requirements.txt
absolufy-imports==0.3.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
audioop-lts==0.2.2
authlib==1.6.3
backoff==2.2.1
bcrypt==4.3.0
brotli==1.1.0
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
chromadb==1.0.21
click==8.2.1
cloudpickle==3.1.1
coloredlogs==15.0.1
cryptography==45.0.7
dataclasses-json==0.6.7
distro==1.9.0
docstring-parser==0.17.0
duckdb==1.3.2
duckduckgo-search==8.1.1
durationpy==0.10
fastapi==0.116.1
ffmpy==0.6.1
filelock==3.13.1
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2024.6.1
google-adk==1.14.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.113.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.37.0
google-cloud-bigtable==2.32.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-spanner==3.57.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-crc32c==1.7.1
google-genai==1.36.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gradio==5.45.0
gradio-client==1.13.0
graphviz==0.21
greenlet==3.2.4
groovy==0.1.2
grpc-google-iam-v1==0.14.2
grpc-interceptor==0.15.4
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
importlib-metadata==8.7.0
importlib-resources==6.5.2
jinja2==3.1.4
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==33.1.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-text-splitters==0.3.11
langsmith==0.4.27
lxml==6.0.1
mako==1.3.10
markdown-it-py==4.0.0
markupsafe==3.0.2
marshmallow==3.26.1
mcp==1.14.0
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.6.4
mypy-extensions==1.1.0
networkx==3.3
numpy==2.3.3
oauthlib==3.3.1
onnxruntime==1.22.1
opentelemetry-api==1.37.0
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.3
overrides==7.7.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
posthog==5.4.0
primp==0.15.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.11.8
pydantic-core==2.33.2
pydantic-settings==2.10.1
pydub==0.25.1
pygments==2.19.2
pyparsing==3.2.4
pypika==0.48.9
pyproject-hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
ruff==0.13.0
sacremoses==0.1.1
safehttpx==0.1.6
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
semantic-version==2.10.0
sentence-transformers==5.1.0
setuptools==70.2.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.43
sqlalchemy-spanner==1.16.0
sqlparse==0.5.3
sse-starlette==3.0.2
starlette==0.47.3
sympy==1.13.3
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.22.0
tomlkit==0.13.3
torch==2.8.0+cpu
tqdm==4.67.1
transformers==4.56.1
typer==0.17.4
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
yarl==1.20.1
zipp==3.23.0
zstandard==0.24.0
